Make confusion matrix an object instead
